+++ lib/vizcore/server.rb
@@ -8,5 +8,6 @@ end
 
 require_relative "server/frame_broadcaster"
 require_relative "server/rack_app"
+require_relative "server/scene_dependency_watcher"
 require_relative "server/runner"
 require_relative "server/websocket_handler"